+++ src/routes/selectBusinessunits.tsx
@@ -4,7 +4,7 @@ import { IBusinessUnit } from "@pages/selectBusinessUnits/outlets/BusinessUnit/t
 import { SelectBusinessUnits } from "@pages/selectBusinessUnits";
 import { BusinessUnits } from "@pages/selectBusinessUnits/outlets/BusinessUnit";
 import { ErrorPage } from "@design/layout/ErrorPage";
-import { ErrorNotBusinessUnit } from "@pages/errors/ErrorNotBusinessUnit";
+import { ErrorNotBusinessUnit } from "@pages/errors/errorNotBusinessUnit";
 import { AuthAndData } from "@context/authAndDataProvider";
 import { CheckingCredentials } from "@pages/login/checkingCredentials";
 import { Loading } from "@pages/login/loading";
